// src/app/attribute/attribute.enum.ts
export enum AttributeValueTypes {
  TEXT = 'text',
  TEXT_LONG = 'text-long',
  DATE = 'date',
  DATE_TIME = 'date-time',
  SINGLE_SELECT = 'single-select',
  MULTI_SELECT = 'multi-select',
  HTML = 'html',
  NUMBER = 'number',
}

export enum AttributeValidationTypes {
  POSITIVE = 'positive',
  EMAIL = 'email',
  PHONE_NUMBER = 'phone-number',
}